Issue: Added an api view for user registration
