+++ slpnotify/urls.py
@@ -26,7 +26,7 @@ from drf_yasg.views import get_schema_view
 from drf_yasg import openapi
 
 from main.urls import urlpatterns as main_urls
-from main.views import SignIn, SignOut, SetAddressView
+from main import views
 
 schema_view = get_schema_view(
    openapi.Info(
@@ -41,9 +41,10 @@ schema_view = get_schema_view(
 
 
 urlpatterns = [
-    path('signin/', SignIn.as_view(), name='signin'),
-    path('signout/', SignOut.as_view(), name='signout'),
-    path('set-address/', SetAddressView.as_view(), name='setaddress'),
+    path('signin/', views.SignIn.as_view(), name='signin'),
+    path('signup/', views.SignUp.as_view(), name='signup'),
+    path('signout/', views.SignOut.as_view(), name='signout'),
+    path('set-address/', views.SetAddressView.as_view(), name='setaddress'),
     path('accounts/login/', admin.site.urls),
     path('api/v1/', include(main_urls)),
     url(r'^api/swagger(?P<format>\.json|\.yaml)$', schema_view.without_ui(cache_timeout=0), name='schema-json'),
